move DataProcessingError in separate file
